Refactor, create service-account-utils
