+++ packages/celotool/src/lib/aks-prometheus.ts
@@ -2,7 +2,12 @@ import fs from 'fs'
 import { createNamespaceIfNotExists } from './cluster'
 import { envVar, fetchEnv } from './env-utils'
 import { installGenericHelmChart, removeGenericHelmChart } from './helm_deploy'
-import { execCmdWithExitOnFailure, outputIncludes } from './utils'
+import {
+  createServiceAccountIfNotExists,
+  getServiceAccountEmail,
+  getServiceAccountKey,
+} from './service-account-utils'
+import { execCmdWithExitOnFailure } from './utils'
 
 const helmChartPath = '../helm-charts/prometheus-stackdriver'
 const releaseName = 'prometheus-stackdriver'
@@ -44,9 +49,7 @@ async function getPrometheusGcloudServiceAccountKeyBase64(kubeClusterName: strin
 
   const serviceAccountEmail = await getServiceAccountEmail(serviceAccountName)
   const serviceAccountKeyPath = `/tmp/gcloud-key-${serviceAccountName}.json`
-  await execCmdWithExitOnFailure(
-    `gcloud iam service-accounts keys create ${serviceAccountKeyPath} --iam-account ${serviceAccountEmail}`
-  )
+  await getServiceAccountKey(serviceAccountEmail, serviceAccountKeyPath)
   return fs.readFileSync(serviceAccountKeyPath).toString('base64')
 }
 
@@ -55,30 +58,15 @@ async function getPrometheusGcloudServiceAccountKeyBase64(kubeClusterName: strin
 async function createPrometheusGcloudServiceAccount(serviceAccountName: string) {
   const gcloudProjectName = fetchEnv(envVar.TESTNET_PROJECT_NAME)
   await execCmdWithExitOnFailure(`gcloud config set project ${gcloudProjectName}`)
-  const serviceAccountExists = await outputIncludes(
-    `gcloud iam service-accounts list --filter=NAME=${serviceAccountName}`,
-    serviceAccountName,
-    `Service account ${serviceAccountName} exists`
-  )
-  if (serviceAccountExists) {
-    return
+  const accountCreated = await createServiceAccountIfNotExists(serviceAccountName)
+  if (accountCreated) {
+    const serviceAccountEmail = await getServiceAccountEmail(serviceAccountName)
+    await execCmdWithExitOnFailure(
+      `gcloud projects add-iam-policy-binding ${gcloudProjectName} --role roles/monitoring.metricWriter --member serviceAccount:${serviceAccountEmail}`
+    )
   }
-  await execCmdWithExitOnFailure(
-    `gcloud iam service-accounts create ${serviceAccountName} --display-name ${serviceAccountName}`
-  )
-  const serviceAccountEmail = await getServiceAccountEmail(serviceAccountName)
-  await execCmdWithExitOnFailure(
-    `gcloud projects add-iam-policy-binding ${gcloudProjectName} --role roles/monitoring.metricWriter --member serviceAccount:${serviceAccountEmail}`
-  )
 }
 
 function getServiceAccountName(kubeClusterName: string) {
   return `prometheus-aks-${kubeClusterName}`
-}
-
-async function getServiceAccountEmail(serviceAccountName: string) {
-  const [output] = await execCmdWithExitOnFailure(
-    `gcloud iam service-accounts list --filter="displayName:${serviceAccountName}" --format='value[terminator=""](email)'`
-  )
-  return output
 }
