+++ packages/celotool/src/lib/service-account-utils.ts
@@ -1,0 +1,33 @@
+import { execCmdWithExitOnFailure, outputIncludes } from './utils'
+
+// createServiceAccountIfNotExists creates a service account with the given name
+// if it does not exist. Returns if the account was created.
+export async function createServiceAccountIfNotExists(name: string) {
+  // TODO: add permissions for cloudsql editor to service account
+  const serviceAccountExists = await outputIncludes(
+    `gcloud iam service-accounts list`,
+    name,
+    `Service account ${name} exists, skipping creation`
+  )
+  if (!serviceAccountExists) {
+    await execCmdWithExitOnFailure(
+      `gcloud iam service-accounts create ${name} --display-name="${name}"`
+    )
+  }
+  return !serviceAccountExists
+}
+
+// getServiceAccountEmail returns the email of the service account with the
+// given name
+export async function getServiceAccountEmail(serviceAccountName: string) {
+  const [output] = await execCmdWithExitOnFailure(
+    `gcloud iam service-accounts list --filter="displayName:${serviceAccountName}" --format='value[terminator=""](email)'`
+  )
+  return output
+}
+
+export function getServiceAccountKey(serviceAccountEmail: string, keyPath: string) {
+  return execCmdWithExitOnFailure(
+    `gcloud iam service-accounts keys create ${keyPath} --iam-account ${serviceAccountEmail}`
+  )
+}
